requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
pyodbc==4.0.32
python-dotenv==0.19.2
SQLAlchemy==1.4.32
Werkzeug==2.0.3
